♻️ refactor: extract server call llm context build
